src/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.40.0",
    "flask-socketio>=5.3.7",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.0.3",
    "openai>=1.50.0",
    "paho-mqtt>=2.1.0",
    "pvporcupine>=3.0.3",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "spotipy>=2.24.0",
]
